requirements.txt
python-telegram-bot==13.7
google-cloud-speech==2.21.0
google-cloud-language==2.11.1
firebase-admin==6.2.0
python-dotenv==1.0.0
